refactor home route
