+++ src/views/onboarding/components/ToolDownload.vue
@@ -67,11 +67,18 @@ export default class ToolDownload extends Vue {
 
 <style scoped>
 .pkg-progress {
-  margin-top: 3%;
+  margin-top: 1%;
+  display: flex;
+  justify-content: center;
+  flex-direction: row;
+  width: 100%;
+  align-items: center;
+  align-self: center;
 }
 
 .progressText {
-  float: left;
-  width: 50%;
+  width: 30%;
+  display: flex;
+  flex-direction: column;
 }
 </style>
