+++ src/views/onboarding/components/ToolsDownloadStep.vue
@@ -1,27 +1,30 @@
 <template>
-  <div>
+  <div class="tools-download-step">
     <i class="arrow go-back right" v-on:click="selectToolsSetup('empty')"></i>
     <h4>ESP-IDF Tools</h4>
     <p>Define ESP-IDF tools install directory.</p>
-    <input type="text" class="text-size" v-model="idfTools" />
-    <font-awesome-icon
-      :icon="folderIcon"
-      class="open-icon"
-      @mouseover="folderIcon = 'folder-open'"
-      @mouseout="folderIcon = 'folder'"
-      v-on:click="openFolder"
-    />
-    <br />
-    <button v-on:click.once="downloadTools" class="onboarding-button">
-      Download
-    </button>
-    <button
-      v-on:click="selectToolsSetup('manual')"
-      class="onboarding-button"
-      v-if="isInstallationCompleted && isPyInstallCompleted"
-    >
-      Go to next step
-    </button>
+    <div class="tools-input">
+      <input type="text" class="text-size" v-model="idfTools" />
+      <font-awesome-icon
+        :icon="folderIcon"
+        class="open-icon"
+        @mouseover="folderIcon = 'folder-open'"
+        @mouseout="folderIcon = 'folder'"
+        v-on:click="openFolder"
+      />
+    </div>
+    <div class="tools-input">
+      <button v-on:click.once="downloadTools" class="onboarding-button">
+        Download
+      </button>
+      <button
+        v-on:click="selectToolsSetup('manual')"
+        class="onboarding-button"
+        v-if="isInstallationCompleted && isPyInstallCompleted"
+      >
+        Go to next step
+      </button>
+    </div>
     <ToolDownload
       :tool="toolVersion"
       v-for="toolVersion in requiredToolsVersions"
@@ -67,4 +70,15 @@ export default class ToolsDownloadStep extends Vue {
 }
 </script>
 
-<style></style>
+<style>
+.tools-download-step {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+}
+.tools-input {
+  display: flex;
+  width: 90%;
+  justify-content: center;
+}
+</style>
